week2/doctorbooking/doctorbookingapi/requirements.txt
alembic==1.16.1
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
defusedxml==0.7.1
Flask==3.1.1
flask-cors==6.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
httpie==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.4
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
PySocks==1.7.1
python-dotenv==1.1.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
setuptools==80.8.0
SQLAlchemy==2.0.41
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
